[PRMT-229] - Create performance_run_bulk_upload.py script
