+++ .github/workflows/performance-base-run-bulk-upload.yml
@@ -85,7 +85,7 @@ jobs:
 
       - name: Run Bulk Upload
         run: |
-          python run_bulk_upload.py \
+          python performance_run_bulk_upload.py \
             --environment "${{ inputs.sandbox }}" \
             --start-bulk-upload
         working-directory: ./tests/bulk-upload/scripts
